Create UI_folder_file_sheet_filter.py
This code creates UI asking for directory by the user and then drops down the list of excel files in it. Post selecting the file, list of sheets in it pops up, needs to select that and later, that particular sheet from the file is read and necessary filtering is done.